backend/requirements.txt
asgiref==3.5.0
cffi==1.15.0
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==3.1.5
django-mongoengine==0.4.6
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
imbalanced-learn==0.9.0
imblearn==0.0
joblib==1.1.0
mongoengine==0.24.1
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
psycopg2==2.9.3
pycparser==2.21
pymongo==4.0.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.2.4
threadpoolctl==3.1.0
tzdata==2022.1
zope.event==4.5.0
zope.interface==5.4.0
